// schemas/public/Menu.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import { z } from 'zod';

/** Identifier type for public.menu */
export type MenuId = string & { __brand: 'MenuId' };

/** Represents the table public.menu */
export default interface Menu {
  id: MenuId;

  ordering: number;

  active_from: Date | null;

  active_to: Date | null;

  file_data: string | null;

  file_name: string | null;

  created_at: Date;
}

/** Represents the initializer for the table public.menu */
export interface MenuInitializer {
  /** Default value: uuid_generate_v4() */
  id?: MenuId;

  ordering: number;

  active_from?: Date | null;

  active_to?: Date | null;

  file_data?: string | null;

  file_name?: string | null;

  /** Default value: CURRENT_DATE */
  created_at?: Date;
}

/** Represents the mutator for the table public.menu */
export interface MenuMutator {
  id?: MenuId;

  ordering?: number;

  active_from?: Date | null;

  active_to?: Date | null;

  file_data?: string | null;

  file_name?: string | null;

  created_at?: Date;
}

export const menuId = z.string() as unknown as z.Schema<MenuId>;

export const menu = z.object({
  id: menuId,
  ordering: z.number(),
  active_from: z.date().nullable(),
  active_to: z.date().nullable(),
  file_data: z.string().nullable(),
  file_name: z.string().nullable(),
  created_at: z.date(),
}) as unknown as z.Schema<Menu>;

export const menuInitializer = z.object({
  id: menuId.optional(),
  ordering: z.number(),
  active_from: z.date().optional().nullable(),
  active_to: z.date().optional().nullable(),
  file_data: z.string().optional().nullable(),
  file_name: z.string().optional().nullable(),
  created_at: z.date().optional(),
}) as unknown as z.Schema<MenuInitializer>;

export const menuMutator = z.object({
  id: menuId.optional(),
  ordering: z.number().optional(),
  active_from: z.date().optional().nullable(),
  active_to: z.date().optional().nullable(),
  file_data: z.string().optional().nullable(),
  file_name: z.string().optional().nullable(),
  created_at: z.date().optional(),
}) as unknown as z.Schema<MenuMutator>;